✨ Give collaboration its own settings page
The feature is now called Collaboration: sharing collided with the
product's core verb (people share polls with participants), while
collaboration names the mode of working and matches the pay wall's team
collaboration benefit. A dedicated page in the settings sidebar is
discoverable by name, keeps general settings to identity and danger
zone, and gives future membership semantics settings a home. The row is
labeled Work together with the access consequence in its description, so
page, row and description each add information without repeating.
