+++ apps/web/public/locales/en/app.json
@@ -138,6 +138,8 @@
   "clockPreferences": "Clock preferences",
   "clockPreferencesDescription": "Set your preferred time zone and time format.",
   "closePoll": "Close",
+  "collaboration": "Collaboration",
+  "collaborationDescription": "How members work in this space",
   "colors": "Colors",
   "colorsDescription": "Primary colors used for theming",
   "commandMenu": "Command menu",
@@ -493,8 +495,8 @@
   "membersInactiveAlertTitle": "Members are inactive",
   "membersSettingsDescription": "Manage space members, invite new users, and control access permissions.",
   "membersSharingModeChange": "Change",
-  "membersSharingModeIndependently": "Members work independently in this space and only see what they create themselves.",
-  "membersSharingModeTogether": "Members work together in this space and can see and manage everything created here.",
+  "membersWorkIndependently": "Members work independently in this space and only see what they create themselves.",
+  "membersWorkTogether": "Members work together in this space and have full access to everything created here.",
   "microsoft365": "Microsoft 365",
   "minutesAbbreviation": "min",
   "mixedDurations": "Mixed",
@@ -773,10 +775,6 @@
   "spaces": "Spaces",
   "spacesDescription": "Manage the spaces you belong to",
   "spaceSettings": "Space settings",
-  "spaceSharedHint": "Members can see and manage everything created in this space.",
-  "spaceSharedLabel": "Shared space",
-  "spaceSharingDescription": "How members work in this space",
-  "spaceSharingTitle": "Sharing",
   "spaceTypeLabel": "What will you be using it for?",
   "spaceTypePersonal": "Personal",
   "spaceTypeWork": "Work",
@@ -874,6 +872,8 @@
   "weekView": "Week view",
   "whiteLabelAddon": "White Label Add-on",
   "whiteLabelAddonDescription": "This add-on allows you to brand Rallly with your own logo and colors.",
+  "workTogether": "Work together",
+  "workTogetherDescription": "Give members full access to everything created in this space.",
   "wrongVerificationCode": "Your verification code is incorrect",
   "yahoo": "Yahoo",
   "yearlyBilling": "Yearly billing",
