+++ apps/web/src/app/[locale]/(space)/(dashboard)/members/components/collaboration-mode.tsx
@@ -4,10 +4,11 @@ import Link from "next/link";
 import { useSpace } from "@/features/space/client";
 import { Trans } from "@/i18n/client";
 
-// Passive readout of the space's sharing mode. The control itself lives in
-// general settings; this line tells admins what an invitee will see before
-// they invite, and tells members why their lists only show their own items.
-export function SharingMode() {
+// Passive readout of the space's collaboration mode. The control itself
+// lives in the collaboration settings page; this line tells admins what an
+// invitee will get before they invite, and tells members why their lists
+// only show their own items.
+export function CollaborationMode() {
   const { data: space } = useSpace();
   const isAdmin = space.role === "admin";
 
@@ -15,12 +16,12 @@ export function SharingMode() {
     <p className="mt-1 text-muted-foreground text-sm">
       {space.shared ? (
         <Trans
-          i18nKey="membersSharingModeTogether"
-          defaults="Members work together in this space and can see and manage everything created here."
+          i18nKey="membersWorkTogether"
+          defaults="Members work together in this space and have full access to everything created here."
         />
       ) : (
         <Trans
-          i18nKey="membersSharingModeIndependently"
+          i18nKey="membersWorkIndependently"
           defaults="Members work independently in this space and only see what they create themselves."
         />
       )}
@@ -29,7 +30,7 @@ export function SharingMode() {
           {" "}
           <Link
             className="underline hover:text-foreground"
-            href="/settings/general"
+            href="/settings/collaboration"
           >
             <Trans i18nKey="membersSharingModeChange" defaults="Change" />
           </Link>
